Split FIRK methods into subpackage
